+++ src/app/navigation/AdminStack.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
 import type { AdminStackParamList } from "../../features/admin/navigation/types";
 
-import { AdminHomeScreen } from "../../features/admin/screens/AdminHomeScreen";
+import { AdminHomeStack } from "../../features/admin/navigation/AdminHomeStack";
 import { MealScheduleScreen } from "../../features/admin/screens/MealScheduleScreen";
 import { MedScheduleScreen } from "../../features/admin/screens/MedScheduleScreen";
 import { LocationSettingsScreen } from "../../features/admin/screens/LocationSettingsScreen";
@@ -12,7 +12,7 @@ const Tab = createBottomTabNavigator<AdminStackParamList>();
 export function AdminStack() {
   return (
     <Tab.Navigator>
-      <Tab.Screen name="AdminHome" component={AdminHomeScreen} options={{ title: "Admin home" }} />
+      <Tab.Screen name="AdminHome" component={AdminHomeStack} options={{ title: "Admin home", headerShown: false }} />
       <Tab.Screen name="MealSchedule" component={MealScheduleScreen} options={{ title: "Meal Schedule" }} />
       <Tab.Screen name="MedSchedule" component={MedScheduleScreen} options={{ title: "Med Schedule" }} />
       <Tab.Screen name="LocationSettings" component={LocationSettingsScreen} options={{ title: "Location Settings" }} />
