+++ src/features/admin/navigation/types.ts
@@ -1,3 +1,8 @@
+export type AdminHomeStackParamList = {
+  AdminHomeMain: undefined;
+  Settings: undefined;
+};
+
 export type AdminStackParamList = {
   AdminHome: undefined;
   MealSchedule: undefined;
