+++ src/features/user/screens/HomeScreen.tsx
@@ -1,13 +1,11 @@
 import { View } from "react-native";
 import { useTheme, Text } from "react-native-paper";
-import { logout } from "../../auth/state/authActions";
 import { useAppMode } from "../../../shared/context/appModeContext";
 import { useAuth } from "../../../shared/hooks/useAuth";
-import { PrimaryButton } from "../../../shared/components/Button/PrimaryButton";
 
 export default function HomeScreen() {
 	const theme = useTheme();
-	const { mode, setMode, resetToModePicker } = useAppMode();
+	const { mode } = useAppMode();
 	const { user } = useAuth();
 	const oletustilaLabel = mode === "admin" ? "Admin" : "Käyttäjä";
 
@@ -43,31 +41,6 @@ export default function HomeScreen() {
 			>
 				Nykyinen oletustila: {oletustilaLabel}
 			</Text>
-
-			<PrimaryButton
-				mode="contained"
-				buttonColor={theme.colors.secondary}
-				textColor={theme.colors.onSecondary}
-				onPress={() => setMode("admin")}
-			>
-				Siirry ADMIN näkymään
-			</PrimaryButton>
-			<PrimaryButton
-				mode="outlined"
-				buttonColor={theme.colors.error}
-				textColor={theme.colors.onError}
-				onPress={resetToModePicker}
-			>
-				Tyhjennä oletustila (poista tallennettu valinta)
-			</PrimaryButton>
-			<PrimaryButton
-				mode="contained"
-				buttonColor={theme.colors.secondary}
-				textColor={theme.colors.onSecondary}
-				onPress={logout}
-			>
-				Kirjaudu ulos tililtä
-			</PrimaryButton>
 		</View>
 	);
 }
